created candidate utils
