+++ fapi/main.py
@@ -4,15 +4,24 @@ from  fapi.db import (
       fetch_sessions_by_type,fetch_candidates, fetch_types, insert_login_history, insert_user, get_user_by_username, update_login_info, verify_md5_hash,
 
     fetch_keyword_recordings, fetch_keyword_presentation,fetch_interviews_by_name,insert_interview,delete_interview,update_interview,
- fetch_course_batches, fetch_subject_batch_recording, user_contact, course_content, fetch_candidate_id_by_email,get_candidates_by_status,fetch_interview_by_id,
-    unsubscribe_user, update_user_password ,get_user_by_username, update_user_password ,insert_user,get_google_user_by_email,insert_google_user_db,fetch_candidate_id_by_email,insert_vendor ,fetch_recent_placements , fetch_recent_interviews, get_candidate_by_name, get_candidate_by_id, create_candidate, delete_candidate as db_delete_candidate,update_candidate as db_update_candidate,get_all_placements,
+ fetch_course_batches, fetch_subject_batch_recording, user_contact, course_content, fetch_candidate_id_by_email,fetch_interview_by_id,
+    unsubscribe_user, update_user_password ,get_user_by_username, update_user_password ,insert_user,get_google_user_by_email,insert_google_user_db,fetch_candidate_id_by_email,insert_vendor ,fetch_recent_placements , fetch_recent_interviews,get_all_placements,
  
 
     get_placement_by_id,search_placements_by_candidate_name,create_placement,update_placement,delete_placement,unsubscribe_lead_user,insert_lead_new
  
   
 )
-from  fapi.utils import md5_hash, verify_md5_hash, create_reset_token, verify_reset_token
+
+from  fapi.utils.candidate_utils import (
+    get_all_candidates_paginated,
+    get_candidate_by_id,
+    create_candidate,
+    update_candidate as db_update_candidate,
+    delete_candidate as db_delete_candidate,
+)
+
+from  fapi.util import md5_hash, verify_md5_hash, create_reset_token, verify_reset_token
 from  fapi.auth import create_access_token, verify_token, JWTAuthorizationMiddleware, generate_password_reset_token, verify_password_reset_token, get_password_hash ,create_google_access_token,determine_user_role
 from  fapi.contactMailTemplet import ContactMail_HTML_templete
 from  fapi.mail_service import send_reset_password_email ,send_request_demo_emails
@@ -36,7 +45,7 @@ from fapi.models import LeadBase, LeadCreate, Lead
 
 from fapi.db import get_connection
 import fapi.db as leads_db
-from .db import get_all_candidates_paginated
+
 # from fastapi import HTTPException
 # from models import CandidateModel
 # from db import db
