+++ fapi/util.py
